+++ api-veterinaria/app/Http/Controllers/Veterinarie/VeterinarieController.php
@@ -22,6 +22,11 @@ class VeterinarieController extends Controller
      */
     public function index(Request $request)
     {
+        if(!auth('api')->user()->can("list_veterinary")){
+            return response()->json([
+                "message" => 'THIS ACTION IS UNAUTHORIZED'
+            ],403);
+        }
         $search = $request->get("search");
 
         $veterinaries = User::where(DB::raw("users.name || ' ' || COALESCE(users.surname,'') || ' ' || users.email"),"ilike","%".$search."%")
@@ -69,6 +74,11 @@ class VeterinarieController extends Controller
      */
     public function store(Request $request)
     {
+        if(!auth('api')->user()->can("register_veterinary")){
+            return response()->json([
+                "message" => 'THIS ACTION IS UNAUTHORIZED'
+            ],403);
+        }
         $is_user_exists = User::where("email",$request->email)->first();
         if($is_user_exists){
             return response()->json([
@@ -115,6 +125,11 @@ class VeterinarieController extends Controller
      */
     public function show(string $id)
     {
+        if(!auth('api')->user()->can("profile_veterinary")){
+            return response()->json([
+                "message" => 'THIS ACTION IS UNAUTHORIZED'
+            ],403);
+        }
         $veterinarie = User::findOrfail($id);
         return response()->json([
             "veterinarie" => VeterinarieResource::make($veterinarie),
@@ -126,6 +141,11 @@ class VeterinarieController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if(!auth('api')->user()->can("edit_veterinary")){
+            return response()->json([
+                "message" => 'THIS ACTION IS UNAUTHORIZED'
+            ],403);
+        }
         $is_user_exists = User::where("email",$request->email)->where("id","<>",$id)->first();
         if($is_user_exists){
             return response()->json([
@@ -188,6 +208,11 @@ class VeterinarieController extends Controller
      */
     public function destroy(string $id)
     {
+        if(!auth('api')->user()->can("delete_veterinary")){
+            return response()->json([
+                "message" => 'THIS ACTION IS UNAUTHORIZED'
+            ],403);
+        }
         $veterinarie = User::findOrFail($id);
         if($veterinarie->avatar){
             Storage::delete($veterinarie->avatar);
